GMutex: change implementation to clean up header
Don't #include <pthread.h> or try to use any of the types declared
therein from public glib headers (avoiding the problems that goes along
with that).

Use the exact same definition for all types on all platforms.

This means that the implementation has to be pretty much completely
changed.  We use the existing GRecMutex code as a model and apply the
same concept to GMutex, GCond and GRWLock.

We can drop the _INIT macros now -- everything is zero-initialised.

This represents a (rather small) performance regression on POSIX.  We'll
fix that up by adding an optimised case for Linux and another case for
systems where our new ABI happens to be compatible with the pthread ABI
on that system (OpenBSD and FreeBSD are notable examples of this).

https://bugzilla.gnome.org/show_bug.cgi?id=659866